requirements.txt
configargparse
numpy
tensorboardx
tqdm
opencv-python
python-dateutil
pillow
nerfacc
lpips
scipy
xatlas
trimesh
PyMCubes